+++ Sources/AnyLint/Lint.swift
@@ -127,11 +127,8 @@ public enum Lint {
     }
 
     /// Logs the summary of all detected violations and exits successfully on no violations or with a failure, if any violations.
-    public static func logSummaryAndExit(
-        failOnWarnings: Bool = false,
-        arguments: [String] = [],
-        afterPerformingChecks checksToPerform: () throws -> Void = {}
-    ) throws {
+    public static func logSummaryAndExit(arguments: [String] = [], afterPerformingChecks checksToPerform: () throws -> Void = {}) throws {
+        let failOnWarnings = arguments.contains(Constants.strictArgument)
         let targetIsXcode = arguments.contains(Logger.OutputType.xcode.rawValue)
 
         if targetIsXcode {