+++ CHANGELOG.md
@@ -21,6 +21,8 @@ If needed, pluralize to `Tasks`, `PRs` or `Authors` and list multiple entries se
 ### Added
 - New `-s` / `--strict` option to fail on warnings as well (by default fails only on errors).  
   PR: [#15](https://github.com/Flinesoft/AnyLint/pull/15) | Author: [Cihat Gündüz](https://github.com/Jeehut)
+- New `-l` / `--validate` option to only runs validations for `matchingExamples`, `nonMatchingExamples` and `autoCorrectExamples`.  
+  PR: [#17](https://github.com/Flinesoft/AnyLint/pull/17) | Author: [Cihat Gündüz](https://github.com/Jeehut)
 ### Changed
 - None.
 ### Deprecated